refactor first pass
